Allows user-specific UI features.
TLDR: instead of session.id, the local variable session_id should
now be configured to always be the proper session id (or
technically the empty string for spectators).

To accomplish this, all clients now subscribe to two channels: the
existing game:id channel, as well as a new game:id:user channel,
where user is either 'red', 'blue', or 'spectator'. The
broadcast_to_players utility was added to automatically broadcast
an html update to all of the relevant players, with the proper
session ids.